requirements.txt
albucore==0.0.12
albumentations==1.4.10
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
einops==0.8.0
executing==2.0.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
imageio==2.34.2
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpickle==4.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
more-itertools==10.5.0
mpmath==1.3.0
music21==9.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.18.0
onnxruntime-gpu==1.18.0
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pdf2image==1.17.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.1
psutil==6.0.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
tbb==2021.13.0
threadpoolctl==3.5.0
tifffile==2024.6.18
timm==1.0.7
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typing_extensions==4.12.1
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.11.1
x-transformers==1.30.23
